First buildable version
This is by no means complete, there's a lot to be done on the setup and
configuration side and klipper MCU firmware builds are missing as well.

The rootfs itself should be device independent (although you'd probably
have to pull in machine specific layers), the image recipe is specific
to the ASUS TinkerBoard S.

Overall goal is to be able to build this distro for any target that is
supported by Yocto. The distro should be "single purpose", i.e. no
desktop like Armbian etc, but just provide a minimalistic setup for
printing with a possibility to select and upload your klipper config.

OctoPrint is included but should be optional (i.e. I'd like to be able
to turn it off and just "print from SD" via the local LCD menu.

TODOs of the top of my head:
* add MCU firmware builds (probably via a multiconfig setup)
* add lighttpd with a central entry page and a minimalistic web UI

For the web UI:
* add a possibility to select existing klipper configs and to upload custom
  klipper configs
* add an option to enable/disable OctoPrint
* backup/restore configuration

To build for the tinker board:

$ source setup.sh targets/tinker-board-s/
$ bitbake klipper-image-tinker-board-s

Flashable wic/wic.bmap (use bmaptool) will be located in:

klipper-linux-build/build/deploy/images/tinker-board-s