Use runit instead of init for important services, drop users and groups
using separate user/group for octoprint turned to be a major PITA - when
you tried to make a backup of the octoprint configuration via the
octoprint UI, it seems that it was trying to move the whole
/var/lib/octoprint directory which of course did not work due to
permissions.

Anyway, this is not a multiuser/multipurpose server, not worth to
mess around with users.

There is one user which can be used to log in and su to root:
login: printer
password: printer

Same credentials are set for the web server.

Octoprint sits behind lighttpd, i.e. https://ip/octoprint/ will
internally proxy to the octoprint server which only listens on localhost.

TODO:
* do not start klipper/octoprint by default yet, the web UI should
  guide the user through an initial setup; octoprint should be optional
  anyway.

* root ssh login will be disabled later

* first time dhparam key generation takes way too long

* web UI config wizard

* build MCU firmware